Made changes to check_valid_move to account for player collission.
This was accomplished by adding a for loop that iterated through the vector
of characters and compared their positions to the current position, and
returned if another character was found at the position. This has led to
issues in creating the valid board. I also made the necessary modifications
to all other calls of this function in other classes.